ref(amplify-overtone): move SendEmailPayload/Result to shared types
Move Lambda payload types from handler.ts to types.ts for cross-package
use. Add EmailTemplateName type. Use EmailTemplateName in template
registry and renderTemplate signature for compile-time safety.
