+++ packages/amplify-overtone/src/email/index.ts
@@ -2,6 +2,9 @@ export { defineEmail } from "./factory.js";
 export type {
     EmailProps,
     EmailResources,
+    EmailTemplateName,
+    SendEmailPayload,
+    SendEmailResult,
     SenderConfig,
     SenderWithEmail,
     SenderWithPrefix,
