+++ packages/amplify-overtone/src/index.ts
@@ -1,6 +1,9 @@
 export type {
     EmailProps,
     EmailResources,
+    EmailTemplateName,
+    SendEmailPayload,
+    SendEmailResult,
     SenderConfig,
     SenderWithEmail,
     SenderWithPrefix,
